fix: category renaming
